requirements.txt
fastapi==0.109.1
uvicorn==0.20.0
websockets==12.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic~=1.10.4
pytest==7.1.3
pytest-asyncio==0.20.3
jinja2==3.1.2
aiofiles==23.1.0
python-socketio==5.8.0
python-engineio==4.4.1
